Rewrite Section 1 (Introduction)
This in order to better highlight that an disruptive maintenance has two
distinct phases (fault detection and reconvergence), and that a
temporary outage will not be over until both of these phases have
completed.

This clarifies that BGP Session Culling remains a highly useful
technique and a BCP even in topologies where the fault detection phase
is nearly non-existent (e.g., where the BGP daemon can rely on signaling
from interface link state monitoring and/or BFD).